Create Python code
In this document, we will create a Python script that performs an internet speed test based on a specified IP address. The script will utilize the speedtest-cli library, which allows us to measure the download and upload speeds of the internet connection. This can be particularly useful for diagnosing network issues or simply checking the performance of your internet service.